Make cmd/podman/shared.GenerateCommand tests Linux-only
They rely on existence of /proc/self/exe .
